RESOLVE PROBLEM IN transction

+++ app/src/main/java/com/example/financetrackerapplication/data/datasource/local/entity/AsetEntity.kt
@@ -10,7 +10,7 @@ data class AsetEntity(
     override val id: Long = 0,
 
     override val name: String, // "Bank BCA", "GoPay", "Tunai"
-    val initialBalance: Double, // Saldo awal saat membuat akun
+    val initialBalance: Long, // Saldo awal saat membuat akun
     override val iconName: String?, // Opsional: nama ikon (misal: logo GoPay)
     val groupAset: String // grup aset, seperti tunai, bank ,dll
 ) : TransOptions
