+++ app/src/main/java/com/example/financetrackerapplication/domain/model/ItemTransaction.kt
@@ -6,14 +6,14 @@ data class ItemTransaction(
     // type = HEADER
     val date: String? = null,
     val day: String? = null,
-    val income: Double? = null,
-    val expense: Double? = null,
+    val income: Long? = null,
+    val expense: Long? = null,
 
     // type = ITEM
     val category: String? = null,
     val catatan: String? = null,
     val aset: String? = null,
-    val amount: Double? = null,
+    val amount: Long? = null,
     val percentage: Double? = null
 ) {
     companion object {
